refactor(core): isolate provider turn runner
